Created separate component for no projects available.
