--- src/Modules/Project/project.interface.ts
@@ -1,6 +1,0 @@
-export interface IProjectFromValues {
-  id?: string;
-  name: string;
-  description: string;
-  account_id: string;
-}
