+++ src/Modules/Project/project.scss
@@ -38,3 +38,8 @@
   font-family: Poppins;
   font-weight: 500;
 }
+
+//project list
+.project-list {
+  display: block;
+}
