+++ src/Shared/SharedComponents/SubHeader/subHeader.scss
@@ -1,3 +1,8 @@
+.sub-header-style {
+  display: 'block';
+  justify-content: 'space-between';
+}
+
 //create button
 .create-button-style {
   border-radius: 0.3125rem;
