+++ src/root/withLayout.scss
@@ -1,5 +1,5 @@
 .wrapped-component-style {
-  height: 100vh; /* This sets the height to 100% of the viewport height */
+  height: 100vh;
   overflow: auto;
   padding: 1.3rem;
   box-sizing: border-box;
